kill button in step 4, just above the NodeTree. if no workchain is selected, the button is not shown.
