+++ src/aiidalab_qe/app/result/__init__.py
@@ -2,6 +2,7 @@ import ipywidgets as ipw
 import traitlets as tl
 from aiida import orm
 from aiida.engine import ProcessState
+from aiida.engine.processes import control
 from aiidalab_widgets_base import (
     AiidaNodeViewWidget,
     ProcessMonitor,
@@ -40,8 +41,22 @@ class ViewQeAppWorkChainStatusAndResultsStep(ipw.VBox, WizardAppWidgetStep):
             ],
         )
         ipw.dlink((self, "process"), (self.process_monitor, "value"))
+        
+        self.kill_work_chains_button = ipw.Button(
+            description="Kill workchain",
+            tooltip="Kill the below workchain.",
+            button_style="danger",
+            icon="window-close",
+            disabled=True,
+            layout=ipw.Layout(width="120px",height="40px",display="none"),
+        )
+        self.kill_work_chains_button.on_click(self._on_click_kill_work_chain)
+        self.kill_work_chains_box = ipw.HBox(
+            children=[self.kill_work_chains_button],
+            layout=ipw.Layout(justify_content="flex-end")
+            )
 
-        super().__init__([self.process_status], **kwargs)
+        super().__init__([self.kill_work_chains_box,self.process_status], **kwargs)
 
     def can_reset(self):
         "Do not allow reset while process is running."
@@ -53,7 +68,9 @@ class ViewQeAppWorkChainStatusAndResultsStep(ipw.VBox, WizardAppWidgetStep):
     def _update_state(self):
         if self.process is None:
             self.state = self.State.INIT
+            self.kill_work_chains_button.display = "none"
         else:
+            self.kill_work_chains_button.layout.display = "block"
             process = orm.load_node(self.process)
             process_state = process.process_state
             if process_state in (
@@ -62,13 +79,21 @@ class ViewQeAppWorkChainStatusAndResultsStep(ipw.VBox, WizardAppWidgetStep):
                 ProcessState.WAITING,
             ):
                 self.state = self.State.ACTIVE
+                self.kill_work_chains_button.disabled = False
             elif (
                 process_state in (ProcessState.EXCEPTED, ProcessState.KILLED)
                 or process.is_failed
             ):
                 self.state = self.State.FAIL
+                self.kill_work_chains_button.disabled = True
             elif process.is_finished_ok:
                 self.state = self.State.SUCCESS
+                self.kill_work_chains_button.disabled = True
+                
+    def _on_click_kill_work_chain(self, _=None):
+        workchain = [orm.load_node(self.process)]
+        control.kill_processes(workchain)
+        self._update_state()
 
     @tl.observe("process")
     def _observe_process(self, change):
